+++ changelog.md
@@ -2,6 +2,17 @@
 
 ### `accetto/ubuntu-vnc-xfce`, `accetto/ubuntu-vnc-xfce-firefox`
 
+### Version 1.3.0
+
+(April 2018)
+
+- Handling of Firefox profiles has been redesigned
+  - Folder **profile0.default** for the default Firefox profile has been pre-created and initialized with the file **user.js**, which includes the fix of the issue #2 (see version 1.2.0).
+  - The actual Firefox profile is creted on the first Firefox start.
+  - User can modify the **user.js** file and the whole profile.
+  - The backup copy **user.js.txt** is in the folder **/headless/.mozilla/firefox**, so the default profile can be safely deleted.
+- In the images with Firefox is the VNC user (**headless:headless** by default) the owner of the most folders created in his home folder **/headless**.
+
 ### Version 1.2.0
 
 (April 2018)